#199: Move ProfileHeader
