+++ src/components/Dashboard/Profile/sections/ProfileHeader/ProfileHeader.tsx
@@ -1,7 +1,7 @@
 "use client";
 import { ProfileEntityProps } from "@/components/Dashboard/Profile/types/types";
-import { OpportunityHeader } from "../opportunity";
-import { VolunteerHeader } from "../volunteer";
+import { OpportunityHeader } from "./opportunity";
+import { VolunteerHeader } from "./volunteer";
 
 export const ProfileHeader = (props: ProfileEntityProps) => {
   if ("volunteer" in props && props.volunteer) {
