--- src/components/Dashboard/Profile/sections/ProfileHeader/common/index.ts
@@ -1,1 +1,0 @@
-export { ProfileHeader } from "./ProfileHeader";
